requirements.txt
Cython
librosa==0.9.1
matplotlib
numpy
scipy
tensorboard
torch==1.13.1
torchvision
unidecode
pyopenjtalk
protobuf
tqdm
django
djangorestframework
faster-whisper
python-dotenv
openai
httpx
django-async-stream
daphne